latest update

+++ src/controllers/notesController.js
@@ -6,7 +6,7 @@ export const getAllNotes = async (req, res) => {
 
   const skip = (page - 1) * perPage;
 
-  const notesQuery = Note.find();
+  const notesQuery = Note.find({userId: req.user._id});
 
   if (search) {
     notesQuery.where({ $text: { $search: search } });
@@ -34,7 +34,7 @@ export const getAllNotes = async (req, res) => {
 
 export const getNoteById = async (req, res) => {
   const { noteId } = req.params;
-  const noteById = await Note.findById(noteId);
+  const noteById = await Note.findOne({ _id: noteId, userId: req.user._id });
 
 if (!noteById) {
   throw createHttpError(404, "Note not found");
@@ -45,25 +45,25 @@ if (!noteById) {
 
 export const createNote = async (req, res) => {
   const body = req.body;
-  const note = await Note.create(body);
+  const note = await Note.create({ ...body, userId: req.user._id });
   res.status(201).json(note);
 };
 
 export const deleteNote = async (req, res) => {
   const { noteId } = req.params;
-  const note = await Note.findOneAndDelete({ _id: noteId });
+  const note = await Note.findOneAndDelete({ _id: noteId, userId: req.user._id });
 
   if (!note) {
     throw createHttpError(404, 'Note not found');
   };
 
-  res.status(200).json(note);
+  res.status(204).json(note);
 };
 
 export const updateNote = async (req, res) => {
   const { noteId } = req.params;
   const body = req.body;
-  const note = await Note.findOneAndUpdate({ _id: noteId }, body, {returnDocument: 'after',});
+  const note = await Note.findOneAndUpdate({ _id: noteId, userId: req.user._id }, body, {returnDocument: 'after',});
 
     if (!note) {
       throw createHttpError(404, 'Note not found');
